docs/use-cases/bedrock-kb-retrieval-mcp/requirements.txt
aws-cdk-lib>=2.100.0
constructs>=10.0.0
boto3>=1.34.0
